refactor(phase4b): move grid hooks to features/grid-editor/hooks/
Move grid-specific hooks from src/hooks/ to src/features/grid-editor/hooks/
as part of Phase 4 folder structure redesign.

Hooks moved (9 total):
- useGridAxisLabels.ts
- useGridCoords.ts
- useGridFirstUseHints.ts
- useGridNavigation.ts
- useGridResize.ts
- useGridRowColumnSelection.ts
- useGridTemplate.ts
- useGridZoom.ts
- useInteraction.ts

Changes:
- Create hooks/ directory in features/grid-editor/
- Move grid hooks with updated imports
- Convert old hook files to re-export stubs for backward compatibility
- Update hooks/index.ts barrel to re-export from new location
- Update test mock paths for useGridCoords

Note: useInteraction.ts has dependencies on src/hooks/interactions/ which
were not moved as they're shared with other features.
