+++ src/features/grid-editor/hooks/index.ts
@@ -1,0 +1,10 @@
+// Grid editor hooks - barrel export
+export { useGridAxisLabels } from './useGridAxisLabels';
+export { useGridCoords } from './useGridCoords';
+export { useGridFirstUseHints } from './useGridFirstUseHints';
+export { useGridNavigation } from './useGridNavigation';
+export { useGridResize } from './useGridResize';
+export { useGridRowColumnSelection } from './useGridRowColumnSelection';
+export { useGridTemplate } from './useGridTemplate';
+export { useGridZoom } from './useGridZoom';
+export { useInteraction } from './useInteraction';
